+++ lyon/mbmdcc/MBMDCCManager.hh
@@ -53,9 +53,9 @@ public:
   void setExternalTrigger(uint32_t nc);
   uint32_t externalTrigger();
   void reloadCalibCount();
-  void resetTDC(uint8_t b);
-  void resetFSM(uint8_t b);
-  uint32_t busyCount(uint8_t b);
+  void resetTDC(uint32_t b);
+  void resetFSM(uint32_t b);
+  uint32_t busyCount(uint32_t b);
   uint32_t readRegister(uint32_t adr);
   void writeRegister(uint32_t adr,uint32_t val);
   void c_readreg(http_request m);
